svm iterates ... but very slow.
Gu

+++ src/features/RealFeatures.cpp
@@ -11,28 +11,29 @@ CRealFeatures::~CRealFeatures()
 }
 
 /// get feature vector for sample num
-REAL* CRealFeatures::get_feature_vector(int num, int &len, bool &free)
+REAL* CRealFeatures::get_feature_vector(long num, long &len, bool &free)
 {
-	len=num_features ; 
-	assert(num<num_vectors) ;
-	
-	if (feature_matrix)
-	  {
-	    free=false ;
-	    return &feature_matrix[num*num_features];
-	  } 
-	else
-	  {
-	    free=true ;
-	    REAL* feat=compute_feature_vector(num, len) ;
-	    if (preproc)
-	      {
-		REAL* feat2 = preproc->apply_to_feature_vector(feat, len);
-		delete[] feat ;
-		return feat2 ;
-	      }
-	    return feat ;
-	  }
+  //fprintf(stderr, "RealFeat.get_feature_vector(%ld) [%ld];\n", num, num_vectors) ;
+  len=num_features ; 
+  assert(num<num_vectors) ;
+  
+  if (feature_matrix)
+    {
+      free=false ;
+      return &feature_matrix[num*num_features];
+    } 
+  else
+    {
+      free=true ;
+      REAL* feat=compute_feature_vector(num, len) ;
+      if (preproc)
+	{
+	  REAL* feat2 = preproc->apply_to_feature_vector(feat, len);
+	  delete[] feat ;
+	  return feat2 ;
+	}
+      return feat ;
+    }
 }
 
 void CRealFeatures::free_feature_vector(REAL* feat, bool free)
@@ -42,7 +43,7 @@ void CRealFeatures::free_feature_vector(REAL* feat, bool free)
 } 
 
 /// get the pointer to the feature matrix
-REAL* CRealFeatures::get_feature_matrix(int &num_feat, int &num_vec)
+REAL* CRealFeatures::get_feature_matrix(long &num_feat, long &num_vec)
 {
   num_feat=num_features;
   num_vec=num_vectors;
