+++ src/kernel/LinearKernel.h
@@ -35,7 +35,7 @@ class CLinearKernel: public CKernel
   /// compute kernel function for features a and b
   /// idx_{a,b} denote the index of the feature vectors
   /// in the corresponding feature object
-  REAL compute(CFeatures* a, int idx_a, CFeatures* b, int idx_b);
+  REAL compute(CFeatures* a, long idx_a, CFeatures* b, long idx_b);
   /*    compute_kernel*/
 
   void init_rescale(CFeatures* f);
